Move filter.fastParseInt to convert.ToInt
This has been extracted so that it can be used outside of the filter
package.
